model/ktrnh/requirements.txt
transformers
evaluate
huggingface_hub
datasets
wandb
underthesea
streamlit
pyvi
pyyaml